+++ test/storage.effect.test.js
@@ -9,6 +9,18 @@ import { EFFECT_ERROR } from "../src/message";
 const test = ninos(ava);
 const args = f => f.calls.map(c => c.arguments);
 
+test("Awaiting empty storage does nothing", async t => {
+  const s = new Storage();
+  const emit = t.context.spy(s, "emit");
+
+  t.deepEqual(s.runningEffects(), []);
+
+  await s.wait();
+
+  t.deepEqual(s.runningEffects(), []);
+  t.deepEqual(args(emit), []);
+});
+
 test("Sending messages on a Storage should send them to matching effects", t => {
   const s = new Storage();
   const emit = t.context.spy(s, "emit");
@@ -187,9 +199,7 @@ test("Immediate reply from effect", async t => {
   const effect = t.context.stub(() => ({ tag: "the-reply", data: "foo" }));
   const myEffect = {
     effect,
-    subscribe: {
-      "trigger-effect": true,
-    },
+    subscribe: { "trigger-effect": true },
   };
   const s = new Storage();
   const emit = t.context.spy(s, "emit");
@@ -205,10 +215,13 @@ test("Immediate reply from effect", async t => {
     ["messageQueued", { tag: "trigger-effect" }, ["$"]],
     ["messageMatched", { tag: "trigger-effect" }, [], false],
   ]);
+  t.deepEqual(s.runningEffects(), [
+    { name: undefined, source: ["$"], message: { tag: "trigger-effect" } },
+  ]);
 
   await p;
 
-  // TODO: The paths are wrong, the $ should be removed when replying?
+  t.deepEqual(s.runningEffects(), []);
   t.deepEqual(args(emit), [
     ["messageQueued", { tag: "trigger-effect" }, ["$"]],
     ["messageMatched", { tag: "trigger-effect" }, [], false],
@@ -222,9 +235,7 @@ test("Immediate reply from effect with name", async t => {
   const myEffect = {
     effect,
     name: "My Effect",
-    subscribe: {
-      "trigger-effect": true,
-    },
+    subscribe: { "trigger-effect": true },
   };
   const s = new Storage();
   const emit = t.context.spy(s, "emit");
@@ -240,9 +251,13 @@ test("Immediate reply from effect with name", async t => {
     ["messageQueued", { tag: "trigger-effect" }, ["$"]],
     ["messageMatched", { tag: "trigger-effect" }, [], false],
   ]);
+  t.deepEqual(s.runningEffects(), [
+    { name: "My Effect", source: ["$"], message: { tag: "trigger-effect" } },
+  ]);
 
   await p;
 
+  t.deepEqual(s.runningEffects(), []);
   // TODO: The paths are wrong, the $ should be removed when replying?
   t.deepEqual(args(emit), [
     ["messageQueued", { tag: "trigger-effect" }, ["$"]],
@@ -258,9 +273,7 @@ test("Immediate throw in effect", async t => {
   });
   const myEffect = {
     effect,
-    subscribe: {
-      "trigger-effect": true,
-    },
+    subscribe: { "trigger-effect": true },
   };
   const s = new Storage();
   const emit = t.context.spy(s, "emit");
@@ -269,6 +282,7 @@ test("Immediate throw in effect", async t => {
 
   t.throws(() => s.sendMessage({ tag: "trigger-effect" }), { instanceOf: Error, message: "My Effect error" });
 
+  t.deepEqual(s.runningEffects(), []);
   t.deepEqual(args(effect), [
     [{ tag: "trigger-effect" }, ["$"]],
   ]);
@@ -282,9 +296,7 @@ test("Async reply from effect", async t => {
   const effect = t.context.stub(async () => ({ tag: "the-reply", data: "foo" }));
   const myEffect = {
     effect,
-    subscribe: {
-      "trigger-effect": true,
-    },
+    subscribe: { "trigger-effect": true },
   };
   const s = new Storage();
   const emit = t.context.spy(s, "emit");
@@ -300,10 +312,13 @@ test("Async reply from effect", async t => {
     ["messageQueued", { tag: "trigger-effect" }, ["$"]],
     ["messageMatched", { tag: "trigger-effect" }, [], false],
   ]);
+  t.deepEqual(s.runningEffects(), [
+    { name: undefined, source: ["$"], message: { tag: "trigger-effect" } },
+  ]);
 
   await p;
 
-  // TODO: The paths are wrong, the $ should be removed when replying?
+  t.deepEqual(s.runningEffects(), []);
   t.deepEqual(args(emit), [
     ["messageQueued", { tag: "trigger-effect" }, ["$"]],
     ["messageMatched", { tag: "trigger-effect" }, [], false],
@@ -318,9 +333,7 @@ test("Async throw in effect", async t => {
   });
   const myEffect = {
     effect,
-    subscribe: {
-      "trigger-effect": true,
-    },
+    subscribe: { "trigger-effect": true },
   };
   const s = new Storage();
   const emit = t.context.spy(s, "emit");
@@ -336,14 +349,123 @@ test("Async throw in effect", async t => {
     ["messageQueued", { tag: "trigger-effect" }, ["$"]],
     ["messageMatched", { tag: "trigger-effect" }, [], false],
   ]);
+  t.deepEqual(s.runningEffects(), [
+    { name: undefined, source: ["$"], message: { tag: "trigger-effect" } },
+  ]);
 
   await p;
 
-  // TODO: These paths?
+  t.deepEqual(s.runningEffects(), []);
   t.deepEqual(args(emit), [
     ["messageQueued", { tag: "trigger-effect" }, ["$"]],
     ["messageMatched", { tag: "trigger-effect" }, [], false],
     ["messageQueued", { tag: EFFECT_ERROR, error: new Error("My Effect error") }, ["$", "<"]],
     ["unhandledMessage", { tag: EFFECT_ERROR, error: new Error("My Effect error") }, ["$", "<"]],
+  ]);
+});
+
+test("State async effect", async t => {
+  const effect = t.context.stub(async () => ({ tag: "the-reply", data: "foo" }));
+  const myEffect = {
+    effect,
+    subscribe: { "trigger-effect": true },
+  };
+  const init = t.context.stub(() => updateAndSend({ state: "init" }, { tag: "trigger-effect" }));
+  const update = t.context.stub((s, { data }) => updateData({ state: data }));
+  const subscribe = t.context.stub(({ state }) => state === "init" ? { "the-reply": true } : {});
+  const model = { id: "test", init, update, subscribe };
+  const s = new Storage();
+  const emit = t.context.spy(s, "emit");
+
+  s.addEffect(myEffect);
+
+  s.createState(model);
+
+  t.deepEqual(args(effect), [
+    [{ tag: "trigger-effect" }, ["test"]],
+  ]);
+  t.deepEqual(args(emit), [
+    ["stateCreated", ["test"], undefined, { state: "init" }],
+    ["messageQueued", { tag: "trigger-effect" }, ["test"]],
+    ["messageMatched", { tag: "trigger-effect" }, [], false],
+  ]);
+  t.deepEqual(s.runningEffects(), [
+    { name: undefined, source: ["test"], message: { tag: "trigger-effect" } },
+  ]);
+
+  // TODO: Placeholder for when createState can provide a promise
+  await s.wait();
+
+  t.deepEqual(s.runningEffects(), []);
+  t.deepEqual(args(emit), [
+    ["stateCreated", ["test"], undefined, { state: "init" }],
+    ["messageQueued", { tag: "trigger-effect" }, ["test"]],
+    ["messageMatched", { tag: "trigger-effect" }, [], false],
+    ["messageQueued", { tag: "the-reply", data: "foo" }, ["test", "<"]],
+    ["messageMatched", { tag: "the-reply", data: "foo" }, ["test"], false],
+    ["stateNewData", { state: "foo" }, ["test"], { tag: "the-reply", data: "foo" }],
+  ]);
+});
+
+test("State async effect chain", async t => {
+  const effect = t.context.stub(async () => ({ tag: "the-reply" }));
+  const eff2 = t.context.stub(async () => ({ tag: "finally" }));
+  const myEffect = {
+    effect,
+    subscribe: { "trigger-effect": true },
+  };
+  const mySecondEffect = {
+    effect: eff2,
+    subscribe: { "second-effect": true },
+  };
+  const init = t.context.stub(() => updateAndSend({ state: "init" }, { tag: "trigger-effect" }));
+  const update = t.context.stub((_, { tag }) => {
+    switch (tag) {
+      case "the-reply":
+        return updateAndSend({ state: "updating" }, { tag: "second-effect" });
+      case "finally":
+        return updateData({ state: "done" });
+      default:
+        throw new Error(`Unexpected tag ${tag}`);
+    }
+  });
+  const subscribe = t.context.stub(({ state }) => state === "init" ? { "the-reply": true } : { "finally": true });
+  const model = { id: "test", init, update, subscribe };
+  const s = new Storage();
+  const emit = t.context.spy(s, "emit");
+
+  s.addEffect(myEffect);
+  s.addEffect(mySecondEffect);
+
+  s.createState(model);
+
+  t.deepEqual(args(effect), [
+    [{ tag: "trigger-effect" }, ["test"]],
+  ]);
+  t.deepEqual(args(emit), [
+    ["stateCreated", ["test"], undefined, { state: "init" }],
+    ["messageQueued", { tag: "trigger-effect" }, ["test"]],
+    ["messageMatched", { tag: "trigger-effect" }, [], false],
+  ]);
+  t.deepEqual(s.runningEffects(), [
+    { name: undefined, source: ["test"], message: { tag: "trigger-effect" } },
+  ]);
+
+  // TODO: Await the state creation instead
+  await s.wait();
+
+  t.deepEqual(s.runningEffects(), []);
+  t.deepEqual(args(emit), [
+    ["stateCreated", ["test"], undefined, { state: "init" }],
+    ["messageQueued", { tag: "trigger-effect" }, ["test"]],
+    ["messageMatched", { tag: "trigger-effect" }, [], false],
+    ["messageQueued", { tag: "the-reply" }, ["test", "<"]],
+    ["messageMatched", { tag: "the-reply" }, ["test"], false],
+    ["stateNewData", { state: "updating" }, ["test"], { tag: "the-reply" }],
+    ["messageQueued", { tag: "second-effect" }, ["test"]],
+    ["messageMatched", { tag: "second-effect" }, [], false],
+    ["messageQueued", { tag: "finally" }, ["test", "<"]],
+    ["messageMatched", { tag: "finally" }, ["test"], false],
+    ["stateNewData", { state: "done" }, ["test"], { tag: "finally" }],
   ]);
 });